Add 5th sketch, centralize color pool
