+++ src/main/kotlin/Main.kt
@@ -3,7 +3,8 @@ import java.awt.Color
 
 //fun main() = PApplet.main(Hexbounce1::class.java.name)
 //fun main() = PApplet.main(Hexbounce2::class.java.name)
-fun main() = PApplet.main(Hexbounce4::class.java.name)
+//fun main() = PApplet.main(Hexbounce4::class.java.name)
+fun main() = PApplet.main(Hexbounce5::class.java.name)
 
 fun Int.rgb() : Int {
     return Color(this).rgb
